added to add story (belum bisa upload)

+++ lib/data/model/story_model.dart
@@ -6,8 +6,8 @@ class Story {
   String? description;
   String? photoUrl;
   String? createdAt;
-  int? lat;
-  int? lon;
+  double? lat;
+  double? lon;
 
   Story(
       {this.id,
